Moving Help specific code to help module.
This moves some of the help specific artifacts out of lib.rs and into
the help module.
