+++ argh/src/lib.rs
@@ -265,9 +265,12 @@
 
 #![deny(missing_docs)]
 
+mod help;
+
 use std::str::FromStr;
 
 pub use {
+    crate::help::{Help, HelpSubCommand, HelpSubCommands},
     argh_derive::{FromArgs, Help},
     argh_shared::{HelpFieldKind, HelpFlagInfo, HelpOptionality, HelpPositionalInfo},
 };
@@ -275,13 +278,13 @@ pub use {
 /// Information about a particular command used for output.
 pub type CommandInfo = argh_shared::CommandInfo<'static>;
 
-/// TODO
+/// Information desribing the help for a command.
 pub type HelpInfo = argh_shared::HelpInfo<'static>;
 
-/// TODO
+/// Information desribing the help for a collection of subcommands.
 pub type HelpSubCommandsInfo = argh_shared::HelpSubCommandsInfo<'static>;
 
-/// TODO
+/// Information desribing the help for a single subcommand.
 pub type HelpSubCommandInfo = argh_shared::HelpSubCommandInfo<'static>;
 
 /// Types which can be constructed from a set of commandline arguments.
@@ -612,50 +615,6 @@ pub struct EarlyExit {
     // TODO replace with std::process::ExitCode when stable.
     pub status: Result<(), ()>,
 }
-
-/// TODO
-pub trait Help : FromArgs {
-    /// TODO
-    const HELP_INFO: &'static HelpInfo;
-
-
-            /// Returns a JSON encoded string of the usage information. This is intended to
-    /// create a "machine readable" version of the help text to enable reference
-    /// documentation generation.
-    fn help_json_from_args(strs: &[&str]) -> Result<String, EarlyExit> {
-        Ok(Self::HELP_INFO.help_json_from_args(strs))
-    }
-}
-
-    /// Returns a JSON encoded string of the usage information based on the command line
-    /// found in argv, identical to `::from_env()`. This is intended to
-    /// create a "machine readable" version of the help text to enable reference
-    /// documentation generation.
-   pub fn help_json<T:Help>() -> Result<String, EarlyExit> {
-        let strings: Vec<String> = std::env::args().collect();
-        //let cmd = cmd(&strings[0], &strings[0]);
-        let strs: Vec<&str> = strings.iter().map(|s| s.as_str()).collect();
-        T::help_json_from_args(strs.as_slice())
-
-    }
-/// TODO
-pub trait HelpSubCommands {
-    /// TODO
-    const HELP_INFO: &'static HelpSubCommandsInfo;
-}
-
-/// TODO
-pub trait HelpSubCommand {
-    /// TODO
-    const HELP_INFO: &'static HelpSubCommandInfo;
-}
-
-impl<T: HelpSubCommand> HelpSubCommands for T {
-    /// TODO
-    const HELP_INFO: &'static HelpSubCommandsInfo =
-        &HelpSubCommandsInfo { optional: false, commands: &[<T as HelpSubCommand>::HELP_INFO] };
-}
-
 
 impl From<String> for EarlyExit {
     fn from(err_msg: String) -> Self {
