+++ argh_derive/src/lib.rs
@@ -23,15 +23,16 @@ mod errors;
 mod help;
 mod parse_attrs;
 
+use crate::help::impl_derive_help;
+
 /// Entrypoint for `#[derive(Help)]`.
 #[proc_macro_derive(Help, attributes(argh))]
 pub fn argh_derive_help(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-   let ast = syn::parse_macro_input!(input as syn::DeriveInput);
-   let gen = impl_derive_help(&ast);
-  // let gen = TokenStream::new();
+    let ast = syn::parse_macro_input!(input as syn::DeriveInput);
+    let gen = impl_derive_help(&ast);
+    // let gen = TokenStream::new();
     gen.into()
 }
-
 
 /// Entrypoint for `#[derive(FromArgs)]`.
 #[proc_macro_derive(FromArgs, attributes(argh))]
@@ -39,27 +40,6 @@ pub fn argh_derive(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let ast = syn::parse_macro_input!(input as syn::DeriveInput);
     let gen = impl_from_args(&ast);
     gen.into()
-}
-
-fn impl_derive_help(input: &syn::DeriveInput) -> TokenStream {
-    let errors = &Errors::default();
-    if !input.generics.params.is_empty() {
-        errors.err(
-            &input.generics,
-            "`#![derive(Help)]` cannot be applied to types with generic parameters",
-        );
-    }
-    let type_attrs = &TypeAttrs::parse(errors, input);
-    let mut output_tokens = match &input.data {
-        syn::Data::Struct(ds) => impl_help_from_args_struct(errors, &input.ident, type_attrs, ds),
-        syn::Data::Enum(de) => impl_help_from_args_enum(errors, &input.ident, type_attrs, de),
-        syn::Data::Union(_) => {
-            errors.err(input, "`#[derive(Help)]` cannot be applied to unions");
-            TokenStream::new()
-        }
-    };
-    errors.to_tokens(&mut output_tokens);
-    output_tokens
 }
 
 /// Transform the input into a token stream containing any generated implementations,
@@ -233,56 +213,6 @@ impl<'a> StructField<'a> {
     }
 }
 
-fn impl_help_from_args_struct(
-    errors: &Errors,
-    name: &syn::Ident,
-    type_attrs: &TypeAttrs,
-    ds: &syn::DataStruct, 
-) -> TokenStream {
-
-    let fields = match &ds.fields {
-        syn::Fields::Named(fields) => fields,
-        syn::Fields::Unnamed(_) => {
-            errors.err(
-                &ds.struct_token,
-                "`#![derive(FromArgs)]` is not currently supported on tuple structs",
-            );
-            return TokenStream::new();
-        }
-        syn::Fields::Unit => {
-            errors.err(&ds.struct_token, "#![derive(FromArgs)]` cannot be applied to unit structs");
-            return TokenStream::new();
-        }
-    };
-
-    let fields: Vec<_> = fields
-    .named
-    .iter()
-    .filter_map(|field| {
-        let attrs = FieldAttrs::parse(errors, field);
-        StructField::new(errors, field, attrs)
-    })
-    .collect();
-
-    let impl_span = Span::call_site();
-
-    let help_info = impl_help(errors, type_attrs, &fields);
-
-    let top_or_sub_cmd_help_impl = top_or_sub_cmd_help_impl(errors, name, type_attrs);
-
-    let trait_impl = quote_spanned! { impl_span =>
-        #[automatically_derived]
-        impl argh::Help for #name {
-            const HELP_INFO: &'static argh::HelpInfo = &#help_info;
-        }
-
-        #top_or_sub_cmd_help_impl
-    };
-
-    trait_impl
-
-}
-
 /// Implements `FromArgs` and `TopLevelCommand` or `SubCommand` for a `#[derive(FromArgs)]` struct.
 fn impl_from_args_struct(
     errors: &Errors,
@@ -333,132 +263,11 @@ fn impl_from_args_struct(
 
             #redact_arg_values_method
         }
+
         #top_or_sub_cmd_impl
     };
 
     trait_impl
-}
-
-fn impl_help<'a>(
-    errors: &Errors,
-    type_attrs: &TypeAttrs,
-    fields: &'a [StructField<'a>],
-) -> TokenStream {
-    let mut subcommands_iter =
-        fields.iter().filter(|field| field.kind == FieldKind::SubCommand).fuse();
-
-    let subcommand: Option<&StructField<'_>> = subcommands_iter.next();
-    for dup_subcommand in subcommands_iter {
-        errors.duplicate_attrs("subcommand", subcommand.unwrap().field, dup_subcommand.field);
-    }
-
-    let impl_span = Span::call_site();
-
-    let mut positionals = vec![];
-    let mut flags = vec![];
-
-    for field in fields {
-        let optionality = match field.optionality {
-            Optionality::None => quote! { argh::HelpOptionality::None },
-            Optionality::Defaulted(_) => quote! { argh::HelpOptionality::Optional },
-            Optionality::Optional => quote! { argh::HelpOptionality::Optional },
-            Optionality::Repeating => quote! { argh::HelpOptionality::Repeating },
-        };
-
-        match field.kind {
-            FieldKind::Positional => {
-                let name = field.arg_name();
-
-                let description = if let Some(desc) = &field.attrs.description {
-                    desc.content.value().trim().to_owned()
-                } else {
-                    String::new()
-                };
-
-                positionals.push(quote! {
-                    argh::HelpPositionalInfo {
-                        name: #name,
-                        description: #description,
-                        optionality: #optionality,
-                    }
-                });
-            }
-            FieldKind::Switch | FieldKind::Option => {
-                let short = if let Some(short) = &field.attrs.short {
-                    quote! { Some(#short) }
-                } else {
-                    quote! { None }
-                };
-
-                let long = field.long_name.as_ref().expect("missing long name for option");
-
-                let description = help::require_description(
-                    errors,
-                    field.name.span(),
-                    &field.attrs.description,
-                    "field",
-                );
-
-                let kind = if field.kind == FieldKind::Switch {
-                    quote! {
-                        argh::HelpFieldKind::Switch
-                    }
-                } else {
-                    let arg_name = if let Some(arg_name) = &field.attrs.arg_name {
-                        quote! { #arg_name }
-                    } else {
-                        let arg_name = long.trim_start_matches("--");
-                        quote! { #arg_name }
-                    };
-
-                    quote! {
-                        argh::HelpFieldKind::Option {
-                            arg_name: #arg_name,
-                        }
-                    }
-                };
-
-                flags.push(quote! {
-                    argh::HelpFlagInfo {
-                        short: #short,
-                        long: #long,
-                        description: #description,
-                        optionality: #optionality,
-                        kind: #kind,
-                    }
-                });
-            }
-            FieldKind::SubCommand => {}
-        }
-    }
-
-    let subcommand = if let Some(subcommand) = subcommand {
-        let subcommand_ty = subcommand.ty_without_wrapper;
-        quote! { Some(<#subcommand_ty as argh::HelpSubCommands>::HELP_INFO) }
-    } else {
-        quote! { None }
-    };
-
-    let description =
-        help::require_description(errors, Span::call_site(), &type_attrs.description, "type");
-    let examples = type_attrs.examples.iter().map(|e| quote! { #e });
-    let notes = type_attrs.notes.iter().map(|e| quote! { #e });
-
-    let error_codes = type_attrs.error_codes.iter().map(|(code, text)| {
-        quote! { (#code, #text) }
-    });
-
-    quote_spanned! { impl_span =>
-        argh::HelpInfo {
-            description: #description,
-            examples: &[#( #examples, )*],
-            notes: &[#( #notes, )*],
-            positionals: &[#( #positionals, )*],
-            flags: &[#( #flags, )*],
-            subcommand: #subcommand,
-            error_codes: &[#( #error_codes, )*],
-        }
-    }
 }
 
 fn impl_from_args_struct_from_args<'a>(
@@ -741,26 +550,6 @@ fn ensure_unique_names(errors: &Errors, fields: &[StructField<'_>]) {
     }
 }
 
-fn top_or_sub_cmd_help_impl(errors: &Errors, name: &syn::Ident, type_attrs: &TypeAttrs) -> TokenStream {
-    if type_attrs.is_subcommand.is_some() {
-        let empty_str = syn::LitStr::new("", Span::call_site());
-        let subcommand_name = type_attrs.name.as_ref().unwrap_or_else(|| {
-            errors.err(name, "`#[argh(name = \"...\")]` attribute is required for subcommands");
-            &empty_str
-        });
-        quote! {
-            #[automatically_derived]
-            impl argh::HelpSubCommand for #name {
-                const HELP_INFO: &'static argh::HelpSubCommandInfo = &argh::HelpSubCommandInfo {
-                    name: #subcommand_name,
-                    info: <#name as argh::Help>::HELP_INFO,
-                };
-            }
-        }
-    } else {
-        quote! {}
-    }
-}
 /// Implement `argh::TopLevelCommand` or `argh::SubCommand` as appropriate.
 fn top_or_sub_cmd_impl(errors: &Errors, name: &syn::Ident, type_attrs: &TypeAttrs) -> TokenStream {
     let description =
@@ -1093,62 +882,6 @@ fn ty_inner<'a>(wrapper_names: &[&str], ty: &'a syn::Type) -> Option<&'a syn::Ty
         }
     }
     None
-}
-
-fn impl_help_from_args_enum(
-    errors: &Errors,
-    name: &syn::Ident,
-    type_attrs: &TypeAttrs,
-    de: &syn::DataEnum,
-) -> TokenStream {
-    parse_attrs::check_enum_type_attrs(errors, type_attrs, &de.enum_token.span);
-
-        // An enum variant like `<name>(<ty>)`
-        struct SubCommandVariant<'a> {
-            ty: &'a syn::Type,
-        }
-    
-        let mut dynamic_type_and_variant = None;
-    
-        let variants: Vec<SubCommandVariant<'_>> = de
-            .variants
-            .iter()
-            .filter_map(|variant| {
-                let name = &variant.ident;
-                let ty = enum_only_single_field_unnamed_variants(errors, &variant.fields)?;
-                if parse_attrs::VariantAttrs::parse(errors, variant).is_dynamic.is_some() {
-                    if dynamic_type_and_variant.is_some() {
-                        errors.err(variant, "Only one variant can have the `dynamic` attribute");
-                    }
-                    dynamic_type_and_variant = Some((ty, name));
-                    None
-                } else {
-                    Some(SubCommandVariant { ty })
-                }
-            })
-            .collect();
-    
-        let variant_ty = variants.iter().map(|x| x.ty).collect::<Vec<_>>();
-
-        let dynamic_commands_help = dynamic_type_and_variant.as_ref().map(|(dynamic_type, _)| {
-            quote! {
-                fn dynamic_commands_help() -> &'static [&'static argh::HelpSubCommandsInfo] {
-                    <#dynamic_type as argh::DynamicHelpSubCommand>::commands()
-                }
-            }
-        });
-
-    quote! {
-        impl argh::HelpSubCommands for #name {
-            const HELP_INFO: &'static argh::HelpSubCommandsInfo = &argh::HelpSubCommandsInfo {
-                optional: false,
-                commands: &[#(
-                    <#variant_ty as argh::HelpSubCommand>::HELP_INFO,
-                )*],
-            };
-            #dynamic_commands_help
-        }
-    }
 }
 
 /// Implements `FromArgs` and `SubCommands` for a `#![derive(FromArgs)]` enum.
